+++ app/employee-side/schedule/page.tsx
@@ -1,10 +1,10 @@
 "use client";
-import "../globals.css";
+
 import { useState, useCallback, useEffect } from "react";
 import { useRouter } from "next/navigation";
-import { ViewToggle } from "../schedule/components/ViewToggle";
-import { Navigation } from "../schedule/components/Navigation";
-import { Calendar, CalendarEvent } from "../schedule/components/Calendar";
+import { ViewToggle } from "../../schedule/components/ViewToggle";
+import { Navigation } from "../../schedule/components/Navigation";
+import { Calendar, CalendarEvent } from "../../schedule/components/Calendar";
 
 import { assignmentsApi } from "@/lib/supabase/assignments";
 import { useAuth } from "@/contexts/AuthContext";
@@ -215,8 +215,8 @@ export default function EmplpyeeSchedule(): React.ReactElement {
   );
 
   return (
-    <div className="schedule-container">
-      <div className="flex justify-between">
+    <div className="schedule-container ">
+      <div className="flex justify-between ">
         <div className="flex flex-col gap-2">
           <h2 className="schedule-title text-primary-dark">Schedule</h2>
           <p className="schedule-subtitle text-gray-500">
