+++ app/employee-side/time-off-request/page.tsx
@@ -2,13 +2,13 @@
 
 import { useState, ReactElement } from "react";
 import { useRouter } from "next/navigation";
-import { TimeOffRequestFormData } from "../types";
+import { TimeOffRequestFormData } from "../../types";
 import { timeOffRequestsApi } from "@/lib/supabase/timeOffRequests";
-import { useTutorial } from "../tutorial/TutorialContext";
-import { TutorialHighlight } from "../components/TutorialHighlight";
+import { useTutorial } from "../../tutorial/TutorialContext";
+import { TutorialHighlight } from "../../components/TutorialHighlight";
 import { useAuth } from "@/contexts/AuthContext";
 import { createClient } from "@/lib/supabase/client";
-import ErrorModal from "../components/ErrorModal";
+import ErrorModal from "../../components/ErrorModal";
 import { ValidationError } from "@/lib/formValidation";
 
 export default function TimeOffRequestPage(): ReactElement {
